Added Python callback example
Added an example to show how to invoke Python methods in the callback
functions.

The example loads a Python module with one function, and calls it when
a command is received, passing the event as an argument. It also reads
the return value and prints it back to the event source.

/cc @sergicastro